pyproject.toml
[project]
name = "llm-fn-dd-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "ddtrace>=3.5.1",
    "langchain-core>=0.3.54",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.14",
    "python-dotenv>=1.1.0",
]
